Move rechunk_seconds to Rechunk.align_utc_seconds
